* Removed org.obiba.onyx.xstream.InjectingReflectionProviderWrapper in favor of identical class in obiba-core * ONYX-1051: VariableDataSource no longer implements InitializingBean since it is impossible to respect the contract during XStream desrialization. * Fixed DataValueConverter

+++ onyx-core/src/test/java/org/obiba/onyx/magma/DataValueConverterTest.java
@@ -65,7 +65,7 @@ public class DataValueConverterTest {
     Data data = new Data(DataType.INTEGER, 42);
     Value value = DataValueConverter.dataToValue(data);
     assertThat(value.getValueType(), equalTo((ValueType) IntegerType.get()));
-    assertThat((Integer) value.getValue(), is(42));
+    assertThat((Long) value.getValue(), is(42l));
   }
 
   @Test
